refactor: Webpack for dev taken from cozy-scripts
